Update package.json and rename bundles (#2)
