+++ LICENSE
@@ -1,5 +1,6 @@
 The MIT License (MIT)
 
+Copyright (c) 2025, Developer Express Inc.
 Copyright (c) 2014-2019 Guyon Roche
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
