+++ test/regression/package.json
@@ -4,14 +4,10 @@
   "description": "Test pack to compare against older versions of exceljs.",
   "private": false,
   "license": "MIT",
-  "author": {
-    "name": "Guyon Roche",
-    "email": "cyber.sapiens@hotmail.com",
-    "url": "https://github.com/guyonroche/exceljs"
-  },
+  "author": "Developer Express Inc.",
   "repository": {
     "type": "git",
-    "url": "https://github.com/guyonroche/exceljs.git"
+    "url": "https://github.com/DevExpress/devextreme-exceljs-fork.git"
   },
   "keywords": [
     "xlsx", "json", "csv", "excel", "font", "border", "fill", "number", "format", "number format", "alignment", "office", "spreadsheet", "workbook",
