+++ test/test-cf.js
@@ -1,4 +1,4 @@
-const Excel = require('../lib/exceljs.nodejs.js');
+const Excel = require('../lib/dx-exceljs-fork.nodejs.js');
 const Range = require('../lib/doc/range');
 
 const HrStopwatch = require('./utils/hr-stopwatch');
