+++ test/test-comments.js
@@ -1,4 +1,4 @@
-const Excel = require('../lib/exceljs.nodejs.js');
+const Excel = require('../lib/dx-exceljs-fork.nodejs.js');
 
 const wb = new Excel.Workbook();
 wb.xlsx
