+++ test/test-image-oneCell.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 const HrStopwatch = require('./utils/hr-stopwatch');
 
-const {Workbook} = require('../lib/exceljs.nodejs');
+const {Workbook} = require('../lib/dx-exceljs-fork.nodejs');
 
 const filename = process.argv[2];
