+++ test/test-image-stream-writer.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const Excel = require('../lib/exceljs.nodejs.js');
+const Excel = require('../lib/dx-exceljs-fork.nodejs.js');
 const HrStopwatch = require('./utils/hr-stopwatch');
 
 const [, , filename] = process.argv;
