+++ test/test-protection-spinCount.js
@@ -1,4 +1,4 @@
-const Excel = require('../lib/exceljs.nodejs.js');
+const Excel = require('../lib/dx-exceljs-fork.nodejs.js');
 const HrStopwatch = require('./utils/hr-stopwatch');
 
 const [, , filename, password] = process.argv;
